% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fiche_cq_station.R
\name{generate_fiche_cq_station}
\alias{generate_fiche_cq_station}
\title{Générer une fiche de Contrôle Qualité des données pour une station}
\usage{
generate_fiche_cq_station(
  stations,
  df_data,
  dossier_sortie = getwd(),
  auteur,
  annee_debut,
  annee_fin,
  interactive = FALSE
)
}
\arguments{
\item{stations}{un vecteur de code de station}

\item{df_data}{un dataframe avec les données nécessaires pour la fiche}

\item{dossier_sortie}{le dossier de sortie des fiches html}

\item{auteur}{le nom de l'auteur de la fiche}

\item{annee_debut}{l'année de début des données}

\item{annee_fin}{l'année de fin des données}

\item{interactive}{un booléen si les graphiques sont interactifs, FALSE par défaut}
}
\value{
un ou des fichiers html dans le dossier de sortie spécifié
}
\description{
Générer une fiche de Contrôle Qualité des données pour une station
}
\examples{
\dontrun{
generate_fiche_cq_station(stations = c("03231000_013", "03231000_014"),
dossier_sortie = "D:/fiches_CQ/",
auteur = "Jean Dupont")
}

}
