% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liste_ope_station.R
\name{get_liste_ope_station}
\alias{get_liste_ope_station}
\title{Obtenir la liste des opérations sur une station}
\usage{
get_liste_ope_station(df, code_sta_pp)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}

\item{code_sta_pp}{le numéro 'code station et point de prélèvement'}
}
\value{
un dataframe avec la liste des opérations pour la station
}
\description{
Obtenir la liste des opérations sur une station
}
\examples{
\dontrun{

aspe_table_fiches <-
mef_creer_table_fiches()

get_liste_ope_station(df = aspe_table_fiches, 
code_sta_pp = "03231000_013")
}

}
