% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liste_stations_dpt.R
\name{get_liste_stations_dpt}
\alias{get_liste_stations_dpt}
\title{Obtenir la liste des stations pour un département}
\usage{
get_liste_stations_dpt(df, dpt)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}

\item{dpt}{un vecteur de numéro de département}
}
\value{
un dataframe avec la liste des stations pour le département
}
\description{
Obtenir la liste des stations pour un département
}
\examples{
\dontrun{
aspe_table_fiches <-
mef_creer_table_fiches()

get_liste_stations_dpt(df = aspe_table_fiches,
dpt = c("14", "27"))
}

}
