% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_codes_sandre_pp.R
\name{mef_ajouter_codes_sandre_pp}
\alias{mef_ajouter_codes_sandre_pp}
\title{Ajoute à un dataframe aspe les codes SANDRE et points de prélèvement}
\usage{
mef_ajouter_codes_sandre_pp(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[aspe:mef_creer_passerelle]{aspe::mef_creer_passerelle()}}}
}
\value{
un dataframe avec les infos de coordonnées stations et le code combinant
code sandre et points de prélèvement \code{code_sta_pp}
}
\description{
Ajoute à un dataframe aspe les codes SANDRE et points de prélèvement
}
\examples{
\dontrun{
df <- 
  aspe::mef_creer_passerelle() \%>\% 
  mef_ajouter_codes_sandre_pp(df)
}
 
}
