% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_facies.R
\name{mef_ajouter_facies}
\alias{mef_ajouter_facies}
\title{Ajoute les données de facies à un dataframe}
\usage{
mef_ajouter_facies(df)
}
\arguments{
\item{df}{un dataframe issu de la fonction \code{\link[aspe:mef_creer_passerelle]{aspe::mef_creer_passerelle()}}}
}
\value{
un dataframe avec les données de facies
}
\description{
Ajoute les données de facies à un dataframe
}
\examples{
\dontrun{
df <-
 aspe::mef_creer_passerelle() \%>\%
 mef_ajouter_facies()
 }
}
