% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_habitat.R
\name{mef_ajouter_habitat}
\alias{mef_ajouter_habitat}
\title{Ajoute les données habitats à un dataframe avec les données aspe}
\usage{
mef_ajouter_habitat(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[aspe:mef_creer_passerelle]{aspe::mef_creer_passerelle()}}
ou \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les données habitats
}
\description{
Ajoute les données habitats à un dataframe avec les données aspe
}
\examples{
\dontrun{
df <- 
  aspe::mef_creer_passerelle() \%>\% 
  mef_ajouter_habitat()
}

}
