% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_infos_context.R
\name{mef_ajouter_infos_context}
\alias{mef_ajouter_infos_context}
\title{Ajoute les informations contextuelles issues de ASPE}
\usage{
mef_ajouter_infos_context(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les infos contextuelles sur les opérations de pêche
}
\description{
Ajoute les informations contextuelles issues de ASPE
}
\examples{
\dontrun{
aspe_table_station1 <- 
  mef_creer_table_fiches() \%>\% 
  dplyr::filter(code_sta_pp == "03174000_005") \%>\% 
  dplyr::select(code_sta_pp, sta_libelle_sandre, ope_id, annee) \%>\% 
  mef_ajouter_infos_context() \%>\% 
  unique()
}
}
