% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_creer_table_base_fiches.R
\name{mef_creer_table_base_fiches}
\alias{mef_creer_table_base_fiches}
\title{Création d'un tableau de base avec les infos ASPE pour les fiches}
\usage{
mef_creer_table_base_fiches()
}
\value{
un dataframe avec les informations utiles
}
\description{
La fonction permet de constituer un tableau de base nécessaire pour la création des fiches.
Les données ASPE doivent être chargées dans l'environnement au préalable.
Seules les stations appartenant aux réseaux RCS, RHP et RRP sont conservées.
}
\examples{
\dontrun{
aspe_table_fiches <- 
  mef_creer_table_base_fiches()
}
}
