% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_creer_table_fiches.R
\name{mef_creer_table_fiches}
\alias{mef_creer_table_fiches}
\title{Création d'un tableau de données ASPE pour les fiches de pêche}
\usage{
mef_creer_table_fiches()
}
\value{
un dataframe avec les informations utiles
}
\description{
Création d'un tableau de données ASPE pour les fiches de pêche
}
\examples{
\dontrun{
df <- mef_creer_table_fiches()
}
}
