% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_creer_table_ope_env.R
\name{mef_creer_table_ope_env}
\alias{mef_creer_table_ope_env}
\title{Création d'un tableau avec les données ASPE pour les paramètres de pêche et env sur les opérations}
\usage{
mef_creer_table_ope_env(df)
}
\arguments{
\item{df}{un dataframe avec les infos de peche et env sur les opérations}
}
\value{
un dataframe avec les infos de peche et env sur les opérations
}
\description{
Création d'un tableau avec les données ASPE pour les paramètres de pêche et env sur les opérations
}
\examples{
\dontrun{
aspe_table_fiches_ope_env <- 
  mef_creer_table_base_fiches() \%>\%
  mef_creer_table_ope_env_fiches()
}
}
