% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_tab_peuplement.R
\name{mef_tab_peuplement}
\alias{mef_tab_peuplement}
\title{Créer un tableau peuplement à partir du dataframe pour les fiches stations}
\usage{
mef_tab_peuplement(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les informations pour le peuplement
}
\description{
Créer un tableau peuplement à partir du dataframe pour les fiches stations
}
\examples{
\dontrun{
aspe_table_fiches \%>\%
  dplyr::filter(code_sta_pp == "03231000_013") \%>\% 
  mef_tab_peuplement()
}

}
