% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_tab_proba_esp_ipr.R
\name{mef_tab_proba_esp_ipr}
\alias{mef_tab_proba_esp_ipr}
\title{Créer un tableau avec les probas IPR pour les espèces dans les stations}
\usage{
mef_tab_proba_esp_ipr(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les informations pour les probabilités IPR
}
\description{
Créer un tableau avec les probas IPR pour les espèces dans les stations
}
\examples{
\dontrun{
aspe_table_fiches \%>\% 
  dplyr::filter(code_sta_pp == "03174000_005") \%>\% 
  mef_tab_proba_esp_ipr()
}
}
