% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_tab_protocole_pts.R
\name{mef_tab_protocole_pts}
\alias{mef_tab_protocole_pts}
\title{Mise en forme d'un tableau protocole pêche points pour description opération}
\usage{
mef_tab_protocole_pts(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les informations pour la description du protocole de pêche
}
\description{
Mise en forme d'un tableau protocole pêche points pour description opération
}
\examples{
\dontrun{
aspe_table_fiches \%>\% 
  dplyr::filter(code_sta_pp == "03231000_013") \%>\% 
  mef_tab_protocole_pts()
}

}
