% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_completude_valeur.R
\name{plot_completude_valeur}
\alias{plot_completude_valeur}
\title{Plot le graphique de completude des valeurs pour une station et liste de paramètres}
\usage{
plot_completude_valeur(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}},
avec une liste de paramètres à inspecter}
}
\value{
un graphique ggplot2
}
\description{
Plot le graphique de completude des valeurs pour une station et liste de paramètres
}
\examples{
\dontrun{
aspe_table_fiches \%>\%
  dplyr::filter(code_sta_pp == "03231000_013") \%>\%
  dplyr::select(code_sta_pp, sta_libelle_sandre, ope_id, annee, pre_id,
                odp_longueur,
                odp_largeur_lame_eau,
                profondeur,
                ope_surface_calculee) \%>\%
  plot_completude_valeur()
}

}
