% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_descript_peche_env.R
\name{plot_descript_peche_env}
\alias{plot_descript_peche_env}
\title{Plot descriptif des paramètres de pêche et environnementaux}
\usage{
plot_descript_peche_env(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot descriptif des paramètres de pêche et environnementaux
}
\examples{
\dontrun{
aspe_table_fiches \%>\%
  dplyr::filter(code_sta_pp == "03231000_013") \%>\%
  plot_descript_peche_env()
  }
}
