% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ipr_metrique_station.R
\name{plot_ipr_metrique}
\alias{plot_ipr_metrique}
\title{Plot un graphique avec les métriques IPR théoriques et observés pour une station}
\usage{
plot_ipr_metrique(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_peuplement]{mef_creer_table_peuplement()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot un graphique avec les métriques IPR théoriques et observés pour une station
}
\examples{
\dontrun{
aspe_table_fiches_ipr \%>\% 
dplyr::filter(code_sta_pp == "03231000_013") \%>\% 
plot_ipr_metrique()
}

}
