% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ipr_param_station.R
\name{plot_ipr_param_station}
\alias{plot_ipr_param_station}
\title{Plot le graphique des paramètres IPR (présences et variations)}
\usage{
plot_ipr_param_station(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}
et \code{\link[=mef_tab_proba_esp_ipr]{mef_tab_proba_esp_ipr()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot le graphique des paramètres IPR (présences et variations)
}
\examples{
\dontrun{
aspe_table_fiches_ipr_env \%>\%
  dplyr::filter(code_sta_pp == "03174000_005") \%>\%
  plot_ipr_param_station()
  }
}
