% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ipr_proba_esp_station.R
\name{plot_ipr_proba_esp_station}
\alias{plot_ipr_proba_esp_station}
\title{Plot le graphique des probas des espèces issues du calcul de l'IPR}
\usage{
plot_ipr_proba_esp_station(df)
}
\arguments{
\item{df}{un dataframe issu de la fonction \code{\link[=mef_tab_proba_esp_ipr]{mef_tab_proba_esp_ipr()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot le graphique des probas des espèces issues du calcul de l'IPR
}
\examples{
\dontrun{
aspe_table_fiches \%>\% 
  dplyr::filter(code_sta_pp == "03174000_005") \%>\% 
  mef_tab_proba_esp_ipr() \%>\% 
  plot_ipr_proba_esp_station()
}

}
