% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ipr_station.R
\name{plot_ipr_station}
\alias{plot_ipr_station}
\title{Plot un graphique des notes IPR pour une station}
\usage{
plot_ipr_station(df, df_col_ipr)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_peuplement]{mef_creer_table_peuplement()}}}

\item{df_col_ipr}{un dataframe contenant les classes IPR, à créer avec la fonction \code{\link[=tab_col_class_ipr]{tab_col_class_ipr()}}}
}
\value{
un graphique ggplot2
}
\description{
Le graphique se compose de 2 parties
}
\examples{
\dontrun{

## tableau stations
aspe_table_fiches_ipr <-
  aspe_table_fiches_peuplement \%>\% 
  dplyr::select(code_sta_pp, 
                sta_libelle_sandre, 
                ope_id, 
                annee, 
                ope_date,
                ipr,
                cli_id,
                cli_libelle) \%>\% 
  dplyr::distinct() 

## tableau classes IPR
df_col_ipr <-
  tab_col_class_ipr()

## graphiques
aspe_table_fiches_ipr \%>\% 
  dplyr::filter(code_sta_pp == "03231000_013") \%>\% 
  plot_ipr_station(df = ., 
                   df_col_ipr = df_col_ipr)
                   
}

}
