% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jour_peche.R
\name{plot_jour_peche}
\alias{plot_jour_peche}
\title{Plot jours de pêche}
\usage{
plot_jour_peche(df, jour_julien = F)
}
\arguments{
\item{df}{un dataframe issu de la fonction \code{\link[=get_liste_ope_station]{get_liste_ope_station()}}}

\item{jour_julien}{un booléen pour afficher les jours de l'année et code couleur saison,
par défaut \code{FALSE}}
}
\value{
un graphique ggplot2
}
\description{
Plot jours de pêche
}
\examples{
\dontrun{
stations_sel <-
  get_liste_ope_station(df = aspe_table_fiches,
                        code_sta_pp = c("03231000_013","03174000_005")
  )

plot_jour_peche(stations_sel, jour_julien = F)
}
}
