% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_objectif_reseau.R
\name{plot_objectif_reseau}
\alias{plot_objectif_reseau}
\title{Plot objectifs réseaux pour une station}
\usage{
plot_objectif_reseau(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot objectifs réseaux pour une station
}
\examples{
\dontrun{
aspe_table_station1 <- 
  mef_creer_table_fiches() \%>\% 
  dplyr::filter(code_sta_pp == "03174000_005")

plot_objectif_reseau(aspe_table_station1)
}
}
