% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proto_prospec_passage.R
\name{plot_proto_prospec_passage}
\alias{plot_proto_prospec_passage}
\title{Plot le graphique Protocole(s) + Passage(s) + Point(s) pour une station}
\usage{
plot_proto_prospec_passage(df)
}
\arguments{
\item{df}{un dataframe issu de la fonction \code{\link[=mef_tab_protocole_pts]{mef_tab_protocole_pts()}} issu
de la jointure avec une table générale \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un graphique ggplot2
}
\description{
Plot le graphique Protocole(s) + Passage(s) + Point(s) pour une station
}
\examples{
\dontrun{
aspe_table_fiches \%>\%
  dplyr::filter(code_sta_pp == "03231000_013") \%>\%
  mef_tab_protocole_pts() \%>\% 
  plot_proto_prospec_passage()
}

}
