% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_infos_context.R
\name{select_infos_context}
\alias{select_infos_context}
\title{Selectionne les informations contextuelles pour fiche station}
\usage{
select_infos_context(df)
}
\arguments{
\item{df}{un dataframe aspe issu de la fonction \code{\link[=mef_creer_table_fiches]{mef_creer_table_fiches()}}}
}
\value{
un dataframe avec les informations contextuelles pour la fiche station
}
\description{
Selectionne les informations contextuelles pour fiche station
}
\examples{
\dontrun{
aspe_table_fiches \%>\% 
dplyr::filter(code_sta_pp == "03231000_013") \%>\% 
select_infos_context()
}

}
