% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_col_classe_ipr.R
\name{tab_col_class_ipr}
\alias{tab_col_class_ipr}
\title{Ajoute un tableau couleur pour les classes IPR}
\usage{
tab_col_class_ipr()
}
\value{
un dataframe avec les informations pour les couleurs des classes IPR
}
\description{
Ajoute un tableau couleur pour les classes IPR
}
\examples{
\dontrun{
tab_col_class_ipr()
}

}
