% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fiche_cq_station.R
\name{generate_fiche_cq_station}
\alias{generate_fiche_cq_station}
\title{Générer une fiche de Contrôle Qualité des données pour une station}
\usage{
generate_fiche_cq_station(stations, dossier_sortie = getwd(), auteur)
}
\arguments{
\item{stations}{un vecteur de code de station}

\item{dossier_sortie}{le dossier de sortie des fiches html}

\item{auteur}{le nom de l'auteur de la fiche}
}
\value{
un ou des fichiers html dans le dossier de sortie spécifié
}
\description{
Générer une fiche de Contrôle Qualité des données pour une station
}
\examples{
\dontrun{
generate_fiche_cq_station(stations = c("03231000_013", "03231000_014"),
dossier_sortie = "D:/fiches_CQ/",
auteur = "Jean Dupont")
}

}
